appointment-api/requirements.txt
fastapi
uvicorn
pandas
python-multipart
sqlalchemy
databases[sqlite]
python-jose[cryptography]
passlib[bcrypt]
fastapi-cors